+++ packages/firebase_remote_config/firebase_remote_config/ios/firebase_remote_config/Sources/firebase_remote_config/FirebaseRemoteConfigPlugin.swift
@@ -211,7 +211,10 @@ public class FirebaseRemoteConfigPlugin: NSObject, FlutterPlugin, FlutterStreamH
         return
       }
       if let update {
-        events(Array(update.updatedKeys))
+        let updatedKeys = Array(update.updatedKeys)
+        DispatchQueue.main.async {
+          events(updatedKeys)
+        }
       }
     }
     return nil